+++ packages/radix-ui-themes/src/components/base-button.props.ts
@@ -2,7 +2,7 @@ import { colorProp, highContrastProp, radiusProp } from '../helpers';
 import type { PropDef } from '../helpers';
 
 const sizes = ['1', '2', '3', '4'] as const;
-const variants = ['solid', 'soft', 'ghost', 'outline', 'surface'] as const;
+const variants = ['classic', 'solid', 'soft', 'ghost', 'outline', 'surface'] as const;
 
 const baseButtonPropDefs = {
   size: { type: 'enum', values: sizes, default: '2', responsive: true },